rename spline to interpolation
