+++ CMakeLists.txt
@@ -14,7 +14,7 @@ set(NMTOOLS_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/include)
 set(NMTOOLS_HEADERS
     ${NMTOOLS_INCLUDE_DIR}/nmtools/curvefit/regression.hpp
     ${NMTOOLS_INCLUDE_DIR}/nmtools/curvefit/sinusoid.hpp
-    ${NMTOOLS_INCLUDE_DIR}/nmtools/curvefit/spline.hpp
+    ${NMTOOLS_INCLUDE_DIR}/nmtools/curvefit/interpolation.hpp
     ${NMTOOLS_INCLUDE_DIR}/nmtools/differentiation/finite.hpp
     ${NMTOOLS_INCLUDE_DIR}/nmtools/fourier/series.hpp
     ${NMTOOLS_INCLUDE_DIR}/nmtools/fourier/transform.hpp
