+++ tests/curvefit/interpolation.cpp
